Fix no-setup-in-describe to correctly detect describe calls
There was one call of `astUtils.isDescribe` which has been called
without the second parameter which caused the function to accept
`undefined` as a valid alias for suite function names. Additionally
there are cases where we don’t get a node name for `CallExpressions`
(e.g. `foo()()`). So in such cases the extracted node name is
`undefined` and `isDescribe` will return true due to the above mentioned
bug.

Fixes: #172