Change to use parentheses in names
This way is slightly more explicit and allows top-level to distinguish between the call of the custom function and the call of the function that's returned.
For example, forEach([1, 2, 3]).describe fails the max-top-level-suites rule if it contained more than one nested suite.
Updates tests to match the new setting names.
Add more tests to guard against regressions (some of these were already supported, such as dotted properties, but not explicitly so).